[Core] Config Deployment Improvements - Perspectives
