+++ bundles/AdminBundle/Resources/views/Admin/Index/index.html.twig
@@ -108,7 +108,7 @@
         <div class="bounce3"></div>
     </div>
 </div>
-{% set runtimePerspective = config.getRuntimePerspective(user) %}
+{% set runtimePerspective = perspectiveConfig.getRuntimePerspective(user) %}
 
 <div id="pimcore_sidebar">
     <div id="pimcore_navigation" style="display:none;">
