+++ doc/Development_Documentation/21_Deployment/03_Configuration_Environments.md
@@ -33,6 +33,7 @@ This feature is currently supported by the following configurations:
 - Predefined properties
 - Predefined asset metadata
 - Static Routes
+- Perspectives
 
 
 You can change the write target individually for each type by using environment variables.
@@ -51,4 +52,5 @@ PIMCORE_WRITE_TARGET_WEB_TO_PRINT=settings-store
 PIMCORE_WRITE_TARGET_PREDEFINED_PROPERTIES=settings-store
 PIMCORE_WRITE_TARGET_PREDEFINED_ASSET_METADATA=settings-store
 PIMCORE_WRITE_TARGET_STATICROUTES=settings-store
+PIMCORE_WRITE_TARGET_PERSPECTIVES=settings-store
 ```
